// packages/core/src/api/model/events.d.ts
export declare enum SimpleListener {
    Message = "onMessage",
    AnyMessage = "onAnyMessage",
    MessageDeleted = "onMessageDeleted",
    Ack = "onAck",
    AddedToGroup = "onAddedToGroup",
    ChatDeleted = "onChatDeleted",
    Battery = "onBattery",
    ChatOpened = "onChatOpened",
    IncomingCall = "onIncomingCall",
    CallState = "onCallState",
    GlobalParticipantsChanged = "onGlobalParticipantsChanged",
    GroupApprovalRequest = "onGroupApprovalRequest",
    ChatState = "onChatState",
    Logout = "onLogout",
    Plugged = "onPlugged",
    StateChanged = "onStateChanged",
    Button = "onButton",
    PollVote = "onPollVote",
    Broadcast = "onBroadcast",
    Label = "onLabel",
    Story = "onStory",
    RemovedFromGroup = "onRemovedFromGroup",
    ContactAdded = "onContactAdded",
    Order = "onOrder",
    NewProduct = "onNewProduct",
    Reaction = "onReaction",
    GroupChange = "onGroupChange"
}
//# sourceMappingURL=events.d.ts.map